Modified - move charities file to charities feature
