+++ lib/features/Charities/business/entities/charity_data.dart
