+++ lib/features/Charities/presentation/pages/charity_detail_screen.dart
@@ -3,18 +3,18 @@ import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
 import 'package:intl/intl.dart' as intl;
 import 'package:provider/provider.dart';
-import 'package:tamizshahr/models/charity.dart';
+import 'package:tamizshahr/features/Charities/business/entities/charity.dart';
 import 'package:tamizshahr/models/customer.dart';
 import 'package:tamizshahr/features/customer_feature/presentation/providers/auth.dart';
-import 'package:tamizshahr/provider/charities.dart';
+import 'package:tamizshahr/features/Charities/presentation/providers/charities.dart';
 import 'package:tamizshahr/features/customer_feature/presentation/providers/customer_info.dart';
-import 'package:tamizshahr/screens/donation_screen.dart';
+import 'package:tamizshahr/features/Charities/presentation/pages/donation_screen.dart';
 import 'package:tamizshahr/widgets/custom_dialog_enter.dart';
-import 'package:tamizshahr/widgets/custom_dialog_pay_charity.dart';
+import 'package:tamizshahr/features/Charities/presentation/widgets/custom_dialog_pay_charity.dart';
 import 'package:tamizshahr/core/logic/en_to_ar_number_convertor.dart';
 
-import '../core/theme/app_theme.dart';
-import '../core/widgets/main_drawer.dart';
+import '../../../../core/theme/app_theme.dart';
+import '../../../../core/widgets/main_drawer.dart';
 
 class CharityDetailScreen extends StatefulWidget {
   static const routeName = '/charityDetailScreen';
