+++ lib/features/Charities/presentation/providers/charities.dart
@@ -3,11 +3,11 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:tamizshahr/models/charity.dart';
-import 'package:tamizshahr/provider/charity_main.dart';
+import 'package:tamizshahr/features/Charities/business/entities/charity.dart';
+import 'package:tamizshahr/features/Charities/presentation/providers/charity_main.dart';
 
-import '../models/search_detail.dart';
-import '../core/constants/urls.dart';
+import '../../../../models/search_detail.dart';
+import '../../../../core/constants/urls.dart';
 
 class Charities with ChangeNotifier {
   List<Charity> _charitiesItems = [];
