+++ lib/features/Charities/presentation/widgets/custom_dialog_pay_charity.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart' as intl;
-import 'package:tamizshahr/models/charity.dart';
+import 'package:tamizshahr/features/Charities/business/entities/charity.dart';
 
-import '../core/theme/app_theme.dart';
-import '../core/logic/en_to_ar_number_convertor.dart';
+import '../../../../core/theme/app_theme.dart';
+import '../../../../core/logic/en_to_ar_number_convertor.dart';
 
 class CustomDialogPayCharity extends StatefulWidget {
   final int totalWallet;
