+++ lib/main.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:provider/provider.dart';
-import 'package:tamizshahr/provider/charities.dart';
+import 'package:tamizshahr/features/Charities/presentation/providers/charities.dart';
 import 'package:tamizshahr/provider/clearings.dart';
 import 'package:tamizshahr/features/store_feature/presentation/providers/orders.dart';
-import 'package:tamizshahr/screens/charity_detail_screen.dart';
-import 'package:tamizshahr/screens/charity_screen.dart';
+import 'package:tamizshahr/features/Charities/presentation/pages/charity_detail_screen.dart';
+import 'package:tamizshahr/features/Charities/presentation/pages/charity_screen.dart';
 import 'package:tamizshahr/screens/clear_screen.dart';
 import 'package:tamizshahr/features/waste_feature/collect_detail_screen.dart';
-import 'package:tamizshahr/screens/donation_screen.dart';
+import 'package:tamizshahr/features/Charities/presentation/pages/donation_screen.dart';
 import 'package:tamizshahr/features/store_feature/presentation/screens/orders_screen.dart';
 import 'package:tamizshahr/screens/wallet_screen.dart';
 import 'package:tamizshahr/features/waste_feature/presentation/wastes_screen_animated_list.dart';
